add refactor unit test and config integration tests

+++ src/infra/db/typeorm/setup.js
@@ -1,15 +1,28 @@
 const typeorm = require('typeorm');
+const { resolve } = require('path');
 
-const typeormServer = new typeorm.DataSource({
-  type: 'sqlite',
-  database: 'db.sqlite',
-  synchronize: true,
-  dropSchema: true,
-  entities: [
-    require('./entities/Usuario.entity-typeorm'), 
-    require('./entities/Livro.entity-typeorm'),
-    require('./entities/Emprestimo.entity-typeorm')
-  ],
-});
+let typeormServer;
+
+if(process.env.NODE_ENV === 'test') {
+  typeormServer = new typeorm.DataSource({
+    type: 'sqlite',
+    database: 'db.test.sqlite',
+    synchronize: true,
+    dropSchema: true,
+    entities: [resolve(__dirname, 'entities/*.entity-typeorm.js')],
+  });
+} else if(process.env.NODE_ENV === 'integration') {
+  typeormServer = new typeorm.DataSource({
+    type: 'postgres',
+    host: 'localhost',
+    database: 'biblioteca_test',
+    synchronize: true,
+    username: 'admin',
+    password: 'admin',
+    port: 5432,
+    entities: [resolve(__dirname, 'entities', '*.entity-typeorm.js')],
+  });
+}
+
 
 module.exports = {typeormServer};
